Matter: use singleton data providers
