+++ lib/matter_glue/platform/bsb/CommissionableDataProviderImpl.hpp
@@ -4,17 +4,8 @@
 
 namespace chip {
 namespace DeviceLayer {
-
-class CommissionableDataProviderImpl : public CommissionableDataProvider {
-public:
-    CHIP_ERROR GetSetupDiscriminator(uint16_t& setupDiscriminator) override;
-    CHIP_ERROR SetSetupDiscriminator(uint16_t setupDiscriminator) override;
-    CHIP_ERROR GetSpake2pIterationCount(uint32_t& iterationCount) override;
-    CHIP_ERROR GetSpake2pSalt(MutableByteSpan& saltBuf) override;
-    CHIP_ERROR GetSpake2pVerifier(MutableByteSpan& verifierBuf, size_t& outVerifierLen) override;
-    CHIP_ERROR GetSetupPasscode(uint32_t& setupPasscode) override;
-    CHIP_ERROR SetSetupPasscode(uint32_t setupPasscode) override;
-};
-
+namespace BSB {
+CommissionableDataProvider* GetCommissionableDataProvider(void);
+} // namespace BSB
 } // namespace DeviceLayer
 } // namespace Chip
